+++ spanprofiler/tracer_test.go
@@ -7,13 +7,14 @@ import (
 	"testing"
 
 	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 	"github.com/stretchr/testify/require"
 	"github.com/uber/jaeger-client-go"
 	jaegercfg "github.com/uber/jaeger-client-go/config"
 )
 
 func TestTracer_pprof_labels_propagation(t *testing.T) {
-	tt := initTestTracer(t)
+	tt := initTestTracer(t, 1)
 	defer func() { require.NoError(t, tt.Close()) }()
 
 	t.Run("root span name and ID are propagated as pprof labels", func(t *testing.T) {
@@ -78,7 +79,7 @@ func TestTracer_pprof_labels_propagation(t *testing.T) {
 	})
 }
 
-func initTestTracer(t *testing.T) io.Closer {
+func initTestTracer(t testing.TB, sampleRate float64) io.Closer {
 	t.Helper()
 	// We can't use mock tracer as we actually rely on the
 	// Jaeger tracer implementation details.
@@ -86,7 +87,7 @@ func initTestTracer(t *testing.T) io.Closer {
 		ServiceName: "test",
 		Sampler: &jaegercfg.SamplerConfig{
 			Type:  jaeger.SamplerTypeConst,
-			Param: 1,
+			Param: sampleRate,
 		},
 		Reporter: &jaegercfg.ReporterConfig{
 			LocalAgentHostPort: "127.0.0.100:16686",
@@ -125,4 +126,20 @@ func spanTags(span opentracing.Span) opentracing.Tags {
 		return nil
 	}
 	return s.Tags()
+}
+
+func BenchmarkTracer(b *testing.B) {
+	testTag := opentracing.Tag{Key: string(ext.Component), Value: "test"}
+	run := func(samplingRate float64) func(b *testing.B) {
+		return func(b *testing.B) {
+			tt := initTestTracer(b, samplingRate)
+			defer func() { require.NoError(b, tt.Close()) }()
+
+			for i := 0; i < b.N; i++ {
+				opentracing.StartSpan("foo", ext.RPCServerOption(nil), testTag)
+			}
+		}
+	}
+	b.Run("unsampled", run(0))
+	b.Run("sampled", run(1))
 }